Squashed commit with hyena-api update and compatibility changes
Added local maven repository with hyena-api

pom.xml updated to use local maven repository

Usage of Hyena-Api library

More meaningful messages on errors

Disabled license checking

Compatibility changes with hyena-api

* ColumnId is now long,
* Hyena now returns triples,
* Scan method now requires catalog.

Jar bump and BigInt-Long cast

Changed local repository for git submodule with hyena-api

+++ presto-hyena/src/main/java/co/llective/presto/hyena/HyenaColumnHandle.java
@@ -29,7 +29,7 @@ public class HyenaColumnHandle
 {
     private final String columnName;
     private final Type columnType;
-    private final int ordinalPosition;
+    private final long ordinalPosition;
 
     @JsonCreator
     public HyenaColumnHandle(
@@ -55,7 +55,7 @@ public class HyenaColumnHandle
     }
 
     @JsonProperty
-    public int getOrdinalPosition()
+    public long getOrdinalPosition()
     {
         return ordinalPosition;
     }
